Move files and edit import
